refactor: consolidate api helpers into utils
